Move all base functions of `base.go` to their own package
