+++ format/plain.go
@@ -3,6 +3,7 @@ package format
 import (
 	"fmt"
 
+	"github.com/princjef/gomarkdoc/formatcore"
 	"github.com/princjef/gomarkdoc/lang"
 )
 
@@ -12,25 +13,25 @@ type PlainMarkdown struct{}
 
 // Bold converts the provided text to bold
 func (f *PlainMarkdown) Bold(text string) (string, error) {
-	return Bold(text), nil
+	return formatcore.Bold(text), nil
 }
 
 // CodeBlock wraps the provided code as a code block. The provided language is
 // ignored as it is not supported in plain markdown.
 func (f *PlainMarkdown) CodeBlock(language, code string) (string, error) {
-	return CodeBlock(code), nil
+	return formatcore.CodeBlock(code), nil
 }
 
 // Header converts the provided text into a header of the provided level. The
 // level is expected to be at least 1.
 func (f *PlainMarkdown) Header(level int, text string) (string, error) {
-	return Header(level, Escape(text))
+	return formatcore.Header(level, formatcore.Escape(text))
 }
 
 // RawHeader converts the provided text into a header of the provided level
 // without escaping the header text. The level is expected to be at least 1.
 func (f *PlainMarkdown) RawHeader(level int, text string) (string, error) {
-	return Header(level, text)
+	return formatcore.Header(level, text)
 }
 
 // LocalHref always returns the empty string, as header links are not supported
@@ -47,25 +48,25 @@ func (f *PlainMarkdown) CodeHref(loc lang.Location) (string, error) {
 
 // Link generates a link with the given text and href values.
 func (f *PlainMarkdown) Link(text, href string) (string, error) {
-	return Link(text, href), nil
+	return formatcore.Link(text, href), nil
 }
 
 // ListEntry generates an unordered list entry with the provided text at the
 // provided zero-indexed depth. A depth of 0 is considered the topmost level of
 // list.
 func (f *PlainMarkdown) ListEntry(depth int, text string) (string, error) {
-	return ListEntry(depth, text), nil
+	return formatcore.ListEntry(depth, text), nil
 }
 
 // Accordion generates a collapsible content. Since accordions are not supported
 // by plain markdown, this generates a level 6 header followed by a paragraph.
 func (f *PlainMarkdown) Accordion(title, body string) (string, error) {
-	h, err := Header(6, title)
+	h, err := formatcore.Header(6, title)
 	if err != nil {
 		return "", err
 	}
 
-	return fmt.Sprintf("%s%s", h, Paragraph(body)), nil
+	return fmt.Sprintf("%s%s", h, formatcore.Paragraph(body)), nil
 }
 
 // AccordionHeader generates the header visible when an accordion is collapsed.
@@ -78,7 +79,7 @@ func (f *PlainMarkdown) Accordion(title, body string) (string, error) {
 //
 //	accordion := format.AccordionHeader("Accordion Title") + "Accordion Body" + format.AccordionTerminator()
 func (f *PlainMarkdown) AccordionHeader(title string) (string, error) {
-	return Header(6, title)
+	return formatcore.Header(6, title)
 }
 
 // AccordionTerminator generates the code necessary to terminate an accordion
@@ -91,10 +92,10 @@ func (f *PlainMarkdown) AccordionTerminator() (string, error) {
 
 // Paragraph formats a paragraph with the provided text as the contents.
 func (f *PlainMarkdown) Paragraph(text string) (string, error) {
-	return Paragraph(text), nil
+	return formatcore.Paragraph(text), nil
 }
 
 // Escape escapes special markdown characters from the provided text.
 func (f *PlainMarkdown) Escape(text string) string {
-	return Escape(text)
+	return formatcore.Escape(text)
 }
