+++ formatcore/base.go
@@ -1,4 +1,4 @@
-package format
+package formatcore
 
 import (
 	"errors"
