+++ formatcore/base_test.go
@@ -1,4 +1,4 @@
-package format
+package formatcore
 
 import (
 	"testing"
